+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Next Version
 
+## 1.1.0
+
+#### Added
+
+* Ability to request now playing artwork https://github.com/evandcoleman/node-appletv/pull/36 by @evandcoleman
+* Support for wake, home, and volume commands https://github.com/evandcoleman/node-appletv/pull/38 by @evandcoleman
+
+[Commits](https://github.com/evandcoleman/node-appletv/compare/1.0.11...1.1.0)
+
 ## 1.0.11
 
 #### Internal